captcha-service/requirements.txt
flask==3.0.0
gunicorn==21.2.0
redis==5.0.1
Pillow==10.1.0
pytesseract==0.3.10
easyocr==1.7.0
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
requests==2.31.0
